# backend/src/main/java/com/jiaozhilong/gisagent/solution/SolutionGenerationController.java
package com.jiaozhilong.gisagent.solution;

import com.jiaozhilong.gisagent.common.api.ApiResponse;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

import java.util.UUID;

@RestController
@RequestMapping("/api/v1")
public class SolutionGenerationController {
    private final SolutionGenerationService service;
    public SolutionGenerationController(SolutionGenerationService service) { this.service = service; }

    @PostMapping("/projects/{projectId}/solution-runs")
    @ResponseStatus(HttpStatus.ACCEPTED)
    @PreAuthorize("hasAuthority('proposal:generate')")
    public ApiResponse<SolutionDtos.GenerationRunResponse> start(@PathVariable String projectId, Authentication authentication,
                                                                 @Valid @RequestBody SolutionDtos.GenerationRequest request) {
        return ApiResponse.ok(service.start(projectId, authentication.getName(), request));
    }

    @GetMapping("/solution-runs/{runId}")
    @PreAuthorize("hasAuthority('proposal:view')")
    public ApiResponse<SolutionDtos.GenerationRunResponse> get(@PathVariable UUID runId) { return ApiResponse.ok(service.get(runId)); }
}